Reproducible ETL (#3)
* Add gitignore

* Reproducible ETL

This PR makes the ETL process simpler and completely reproducible.  A few key points:

* Use superdeduper! There are a few side-effects here.
    * The entries table is now `clean.entries` instead of `dedupe.entries`.
    * It now enforces a minimum set of fields required for an entry to get a dedupe_id.  Previously, no entries had a null dedupe_id.
    * It now uses a recursive algorithm for incorporating the known links through the JIMS system.  It also incorporates known links through the `patid` column in the mental health case system.

* There is now a date field in the entries table!  This is crucial for using collate to generate demographic features in a temporally valid manner.  For many of the data sources, this is simple, but since JIMS and MH have a master name table that must get joined against an "events" table of some sort in order to find a possible date at which they would have known the info.  This is done in a fairly rudimentary manner.
    * Note the TODO: only 1/6th of JIMS name records get a date attached with this method! Are there other tables to include?

* This begins a transition towards using a config directory in the repository root in order to specify all the necessary credentials. This solves the issue with hardcoded `/mnt` paths for etl, and I think it's a good solution for Pipeline, too. It is not complete!  After this is merged, we need to get the rest of the pipeline on board with this scheme.

* The etl/README.md documents the general scheme.

+++ etl/data_cleaning/clean_jocojims2bailmstdefinfo.sql
@@ -51,7 +51,7 @@ as (
         zip_code as zip_code,
         prt as prt,
         amt_tendered as amt_tendered,
-        mni_id as mni_no,
+        mni_id::int as mni_no,
         gcvalidity as gcvalidity,
         gcsrctbl as gcsrctbl,
         gcscore as gcscore,
